convert MultiLeaves and best_leave_values from f32 to i32 millipoints
MultiLeaves.leave_values: Vec<f32> -> Vec<i32>.
WorkingBuffer.best_leave_values: Vec<f32> -> Vec<i32>.
PossiblePlacement.best_possible_equity: f32 -> i32.
AdjustLeaveValue callback: Fn(f32)->f32 -> Fn(i32)->i32.
FoundExchangeMove callback takes i32 leave value.

Shadow play equity arithmetic is now pure i32. All intermediate
values accumulate in millipoints without float conversion.

+++ src/main_leave.rs
@@ -1978,7 +1978,7 @@ fn discover_playability<N: kwg::Node + Sync + Send, L: kwg::Node + Sync + Send>(
                                      _idx: i8,
                                      _word: &[u8],
                                      _score: i32| true,
-                                    |leave_value: f32| leave_value,
+                                    |leave_value: i32| leave_value,
                                     |equity: equity::Equity, play: &movegen::Play| {
                                         match equity.cmp(&best_equity_so_far) {
                                             std::cmp::Ordering::Greater => {
